Build in DTAP mode
Based on the URL of the application the DTAP mode is detected. Based on
this mode the enabled setting is retrieved. Applied entity access on the
afterstartup flow (because a lot of other after startup microflows need
this and the whole after startup flow should have this enabled).
Community commons module is needed for the DTAP detection and is added.